Initial commit of project files
This is mostly an empty app with Material-UI and Framer Motion loaded and tested in the app, but at some point I forgot to commit and fleshed out a noscript lander! Everything needs to be stylized and I could add extra animations, but I am very happy with the experience of the noscript page so far.